Word time widget improvement
- Added 30min time difference support

+++ app/src/main/java/com/dinodevs/greatfitwatchface/settings/LoadSettings.java
@@ -55,7 +55,7 @@ public class LoadSettings {
     public int temp_heart_rate;
     public int temp_calories;
     public boolean am_pm_always;
-    public int world_time_zone;
+    public float world_time_zone;
     public Paint mGPaint;
     public List widgets_list;
     public List circle_bars_list;
@@ -483,7 +483,7 @@ public class LoadSettings {
             this.temp_heart_rate = sharedPreferences.getInt( "temp_heart_rate", 0);
             this.temp_calories = sharedPreferences.getInt( "temp_calories", 0);
             this.am_pm_always = sharedPreferences.getBoolean( "am_pm_always", context.getResources().getBoolean(R.bool.am_pm_always));
-            this.world_time_zone = sharedPreferences.getInt( "world_time_zone", -1);
+            this.world_time_zone = sharedPreferences.getFloat( "world_time_zone", -1f);
 
             // Populate color codes
             String[] colorCodes = context.getResources().getStringArray(R.array.color_codes);
